// src/game/tank/utils/painting/types.ts
export enum Resistances {
    SMOKY = 'SMOKY_RESISTANCE',
    FLAMETHROWER = 'FIREBIRD_RESISTANCE',
    RAILGUN = 'RAILGUN_RESISTANCE',
    RAILGUN_XT = 'RAILGUN_RESISTANCE',
    TWINS = 'TWINS_RESISTANCE',
    RICOCHET = 'RICOCHET_RESISTANCE',
    ISIDA = 'ISIS_RESISTANCE',
    HAMMER = 'SHOTGUN_RESISTANCE',
    SHAFT = 'SHAFT_RESISTANCE',
    FREEZE = 'FREEZE_RESISTANCE',
    THUNDER = 'THUNDER_RESISTANCE',
    VULCAN = 'MACHINE_GUN_RESISTANCE',
    ALL = 'ALL_RESISTANCE'
}